Separando alguns componentes, nenhum css ainda feito
