+++ package.json
@@ -29,7 +29,6 @@
   "dependencies": {
     "axios": "^0.19.0",
     "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-router-dom": "^5.1.2"
+    "react-dom": "^16.12.0"
   }
 }
