+++ src/App.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import { HashRouter as Router, Route } from 'react-router-dom'
 
 // Components
 import Container from './components/Container'
@@ -7,14 +6,18 @@ import WeatherContextProvider from './contexts/WeatherContext'
 
 // Global Styles
 import './assets/styles/global.scss'
+import TodayInfo from './components/TodayInfo/TodayInfo'
+import AnotherDayInfo from './components/AnotherDayInfo'
 
 const App = props => {
     return (
-        <Router>
-            <WeatherContextProvider>
-                <Container></Container>
-            </WeatherContextProvider>
-        </Router>
+        <WeatherContextProvider>
+            <Container>
+              <TodayInfo />
+              <AnotherDayInfo day="tomorrow" />
+              <AnotherDayInfo day="afterTomorrow" />
+            </Container>
+        </WeatherContextProvider>
     )
 }
