+++ src/components/AnotherDayInfo/index.js
@@ -1,0 +1,3 @@
+import AnotherDayInfo from './AnotherDayInfo'
+
+export default AnotherDayInfo
