+++ src/components/Container/container.scss
@@ -20,7 +20,7 @@ vermelho 3 => #ff1818 (255,24,24)
   position: fixed;
   width: 100%;
   height: 100%;
-  background-color: #333;
+  background-color: #121212;
 }
 
 .image_container_copyright {
@@ -31,4 +31,11 @@ vermelho 3 => #ff1818 (255,24,24)
   color: rgba(255,255,255,0.3);
   font-size: 12px;
   text-align: right;
+}
+
+.components_list {
+  width: 100%;
+  max-width: 768px;
+  margin: 0 auto;
+  background: rgba(255,255,255,0.5)
 }
