+++ src/components/TodayInfo/index.js
@@ -1,0 +1,3 @@
+import TodayInfo from './TodayInfo'
+
+export default TodayInfo
